+++ src/crack/mod.rs
@@ -26,6 +26,6 @@ SOFTWARE.
 
 pub use result::CrackResult;
 
-mod indices;
+pub(crate) mod indices;
 mod result;
 pub(crate) mod worker_threads;